backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Anthropic Claude (Primary LLM)
anthropic==0.18.1

# Additional LLM Support
openai==1.10.0
google-generativeai==0.3.2

# Hugging Face (Optional - for local models)
transformers==4.37.0
torch==2.1.2
sentencepiece==0.1.99

# Vector Store & Memory
chromadb==0.4.22
redis==5.0.1
sentence-transformers==2.3.1

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.41

# Security & Rate Limiting
slowapi==0.1.9

# Retry Logic
tenacity==8.2.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.0
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# Type Stubs
types-redis==4.6.0.20240106
types-pyyaml==6.0.12.12
